Moved OS and log pages to a blueprint
